added more functins
